+++ src/main/java/com/team/dec051/member/MailSendService.java
@@ -1,0 +1,54 @@
+package com.team.dec051.member;
+
+import java.util.Random;
+
+import javax.mail.MessagingException;
+import javax.mail.internet.MimeMessage;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.mail.javamail.JavaMailSenderImpl;
+import org.springframework.mail.javamail.MimeMessageHelper;
+import org.springframework.stereotype.Component;
+
+@Component
+public class MailSendService {
+	
+	@Autowired
+	private JavaMailSenderImpl mailSender;
+	private int authNumber; 
+	
+		public void makeRandomNumber() {
+			Random r = new Random();
+			int checkNum = r.nextInt(888888) + 111111;
+			System.out.println("인증번호 : " + checkNum);
+			authNumber = checkNum;
+		}
+
+		public String joinEmail(String email) {
+			makeRandomNumber();
+			String setFrom = "pillarjun416@gmail.com";
+			String toMail = email;
+			String title = "회원 가입 인증";
+			String content = 
+				    "인증 번호는 " + authNumber + "입니다.";
+			mailSend(setFrom, toMail, title, content);
+			return Integer.toString(authNumber);
+		}
+		
+		public void mailSend(String setFrom, String toMail, String title, String content) {
+			MimeMessage message = mailSender.createMimeMessage();
+			// true 매개값을 전달하면 multipart 형식의 메세지 전달이 가능.문자 인코딩 설정도 가능하다.
+			try {
+				MimeMessageHelper helper = new MimeMessageHelper(message,true,"utf-8");
+				helper.setFrom(setFrom);
+				helper.setTo(toMail);
+				helper.setSubject(title);
+				helper.setText(content,true);
+				mailSender.send(message);
+			} catch (MessagingException e) {
+				e.printStackTrace();
+			}
+		}
+		
+	
+}
