+++ src/main/java/com/team/dec051/member/MemberDAO.java
@@ -5,73 +5,22 @@ import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
 import java.net.URLEncoder;
 import java.util.List;
-import java.util.Random;
 
-import javax.mail.MessagingException;
-import javax.mail.internet.MimeMessage;
-import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.apache.ibatis.session.SqlSession;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.JavaMailSenderImpl;
-import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.stereotype.Service;
 
 import com.oreilly.servlet.MultipartRequest;
 import com.oreilly.servlet.multipart.DefaultFileRenamePolicy;
+
 
 @Service
 public class MemberDAO {
 
 	@Autowired
 	private SqlSession ss;
-	private JavaMailSender mailSender;
-	private int authNumber; 
-	
-	public void makeRandomNumber() {
-		Random r = new Random();
-		int checkNum = r.nextInt(888888) + 111111;
-		System.out.println("인증번호 : " + checkNum);
-		authNumber = checkNum;
-	}
-	
-	//이메일 양식
-	public String joinEmail(String m_email) {
-		makeRandomNumber();
-		String setFrom = "pillarjun416@gmail.com"; // email-config에 설정한 자신의 이메일 주소를 입력 
-		String toMail = m_email;
-		String title = "회원 가입 인증 이메일 입니다."; 
-		String content = 
-				"### 사이트 발송" + 
-		        "<br>" + 
-				"인증 번호는 " + authNumber + "입니다." + 
-				"<br>" + 
-				"해당 인증번호를 인증번호 확인란에 기입하여 주세요.";
-
-		mailSend(setFrom, toMail, title, content);
-		return Integer.toString(authNumber);
-	}
-			
-	//이메일 전송 메소드
-	public void mailSend(String setFrom, String toMail, String title, String content) {
-		MimeMessage message = mailSender.createMimeMessage();
-		System.out.println("이메일 전송 메소드 2 !!");
-		try {
-			MimeMessageHelper helper = new MimeMessageHelper(message,true,"utf-8");
-			helper.setFrom(setFrom);
-			helper.setTo(toMail);
-			helper.setSubject(title);
-			helper.setText(content,true);
-			mailSender.send(message);
-			System.out.println("이메일 전송 메소드 Try!!");
-		} catch (MessagingException e) {;
-			e.printStackTrace();
-			System.out.println("이메일 전송 메소드 catch!!");
-		}
-	}
 	
 	public void signup(Member m, HttpServletRequest req) {
 		try {
@@ -108,7 +57,6 @@ public class MemberDAO {
 			List<Member> members = ss.getMapper(MemberMapper.class).getMemberById(m);
 			if (members.size() != 0) {
 				Member dbM = members.get(0);
-				// System.out.println(dbM.getm_id());
 				
 				if (dbM.getM_pw().equals(m.getM_pw())) {
 					req.setAttribute("r", "로그인 성공");
