+++ src/main/resources/application.xml
@@ -1,16 +1,15 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://www.springframework.org/schema/beans https://www.springframework.org/schema/beans/spring-beans.xsd">
-	
-	<!-- Root Context: defines shared resources visible to all other web components -->
+  <beans xmlns="http://www.springframework.org/schema/beans"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:context="http://www.springframework.org/schema/c"
+         xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
+  
 
    <!-- 이메일 인증 관련 빈 등록 -->
-   <bean id="mailSender" class="org.springframework.mail.javamail.JavaMailSenderImpl">
+   <bean id="mailSender" class="org.springframework.mail.javamail.JavaMailSenderImpl" >
       <property name="host" value="smtp.gmail.com" />
       <property name="port" value="587" />
-      <property name="username" value="pillarjun416@gmail.com" />
-      <property name="password" value="fssfowhdksowbqqj" />
+      <property name="username" value="pillarjun416" />
+      <property name="password" value="exhuvjgetamwqbnb" />
       <property name="javaMailProperties">
          <props>
                 <prop key="mail.transport.protocol">smtp</prop>
@@ -23,6 +22,7 @@
 				<prop key="mail.smtp.ssl.protocols">TLSv1.2</prop>
             </props>
       </property>
-      
-   </bean>
-</beans>
+	</bean>
+
+
+  </beans>
