+++ src/main/webapp/WEB-INF/views/member/info.jsp
@@ -11,40 +11,45 @@
 		<form action="member.update" method="post" enctype="multipart/form-data">
 			<tr>
 				<td colspan="2">
-					<input value="${sessionScope.loginMember.m_id }" name="m_id" placeholder="ID" autocomplete="off" maxlength="10" class="i1" readonly="readonly">
+					<input value="${sessionScope.loginMember.m_id }" name="m_id" placeholder="ID" autocomplete="off" readonly="readonly">
 				</td>
 
 			</tr>
 			<tr>
 				<td colspan="2">
-					<input name="m_pw" placeholder="Password" autocomplete="off" autofocus="autofocus" maxlength="10" type="password" class="i1">
+					<input name="m_pw" placeholder="Password" autocomplete="off" autofocus="autofocus" type="password">
 				</td>
 			</tr>
 			<tr>
 				<td colspan="2">
-					<input name="m_pwChk"placeholder="Password Check" autocomplete="off" maxlength="10" type="password" class="i1">
+					<input name="m_pwChk"placeholder="Password Check" autocomplete="off" type="password">
 				</td>
 			</tr>
 			<tr>
 				<td colspan="2">
-					<input name="m_name" placeholder="User Name" autocomplete="off" maxlength="4" class="i1">
+					<input value="${sessionScope.loginMember.m_name }"  name="m_name" autocomplete="off">
 				</td>
 			</tr>
+			<tr>
+				<td colspan="2">
+					<input value="${sessionScope.loginMember.m_email }"  name="m_email" autocomplete="off">
+				</td>
+			</tr>		
 			<tr>
 				<td>
 					<div class="fileBox">
 						<input class="uploadName" value="${sessionScope.loginMember.m_photo }" placeholder="Profile Pic">
 						<label for="m_photo">파일찾기</label>					
-						<input id="m_photo" name="m_photo" type="file">
+						<input id="reUploadPhoto" name="m_photo" type="file">
 					</div>
 				</td>
 			</tr>
 			<tr>
 				<td colspan="2">
-					<button>계정 수정</button>
+					<button id="updateMemberBtn">계정 수정</button>
 			</form>
 			<form  action="member.resign">
-					<button id="resignBtn">계정 삭제</button>
+					<button id="deleteMemberBtn">계정 삭제</button>
 				</td>
 			</tr>
 			</form>
